+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Change Log
 <a name="v0.0.69"></a>
 ## Version [v0.0.69](https://github.com/luizfp/PrologWeb/compare/v0.0.68...v0.0.69) (release-date) [unreleased]
 
+### Features
+* Cria validator para o insert do Modelo de Quiz
+
 <a name="v0.0.68"></a>
 ## Version [v0.0.68](https://github.com/luizfp/PrologWeb/compare/v0.0.67...v0.0.68) (2018-07-14)